gen_mapimg/pyproject.toml
[project]
name = "gen-mapimg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.1.2",
    "pillow>=10.4.0",
]
